+++ include/tedpad/modules/gamepadDescription.h
@@ -11,20 +11,20 @@
 
 namespace tedpad {
 	namespace Module {
-		//Structure: Module( gamepadName (?) -> out_*count (4 * 4) -> in_*count (4 * 4) )
+		//Structure: Module( gamepadName (?) -> out_*count (1 * 4) -> in_*count (1 * 4) )
 		struct GamepadBriefDescription : public ModuleBase {
 			PacketModule to_packetModule() const override;
 			void from_packetModule(PacketModule const &p0) override;
 
 			std::string gamepadName;
-			size_t out_digitalCount = 0;
-			size_t out_analogueCount = 0;
-			size_t out_axisCount = 0;
-			size_t out_bufferCount = 0;
-			size_t in_digitalCount = 0;
-			size_t in_analogueCount = 0;
-			size_t in_axisCount = 0;
-			size_t in_bufferCount = 0;
+			uint8_t out_digitalCount = 0;
+			uint8_t out_analogueCount = 0;
+			uint8_t out_axisCount = 0;
+			uint8_t out_bufferCount = 0;
+			uint8_t in_digitalCount = 0;
+			uint8_t in_analogueCount = 0;
+			uint8_t in_axisCount = 0;
+			uint8_t in_bufferCount = 0;
 
 			GamepadBriefDescription();
 		};
@@ -40,8 +40,8 @@ namespace tedpad {
 
 			GamepadFullDescription();
 		private:
-			size_t get_attributeServerOutCount(Attribute::PlaceHolderType const attr_desc) const;
-			size_t get_attributeServerInCount(Attribute::PlaceHolderType const attr_desc) const;
+			uint8_t get_attributeServerOutCount(Attribute::PlaceHolderType const attr_desc) const;
+			uint8_t get_attributeServerInCount(Attribute::PlaceHolderType const attr_desc) const;
 		};
 	}
 }
